for graft miners
he Graft Network is a retail-oriented payment blockchain and a decentralized alternative payment processing network that brings cryptocurrency to the point of sale and beyond. The Graft Community is a large group of investors, traders, crypto miners and entrepreneurs who have taken the initiative to raise funds and promote the opportunities available by using GRFT as an alternative currency.