Delisting CLO Pool as 2miners is not supported anymore by CLO Team
Dear community!

After the recent events and after seeing the way "2Miners" pool handled the situation, we on Callisto noted a lack of cooperation and professionalism from their side.

Therefor 2miners pool is not supported by Callisto Team anymore.

We have taken the decision to stop any kind of relations with them and not giving support to miners mining on 2Miners pool. We are not going to help or answer any questions about that pool. In case of losses due to attacks or instability, miners who are mining on 2Miners will not be compensated. Callisto Team is not responsible for any kind of problems miners may experience, caused by mining on this pool.

Any problem that miners might have with payments, orphan blocks or losses on "2miners" pool will not receive any kind of support from Callisto Network.

There are plenty of other mining pools that support CLO and have an excellent site, interface and support team. We strongly recommend that you move your hashing powers there. A list of all supported CLO mining pools can be found on our website here:

https://callisto.network/#mining

We will be supporting these pools as we have always done. We will take care of any problems that miners could get mining on the network. We will refund in case of losses for problems on the chain.

Callisto Network.